% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVAFunctions.R
\name{adjust_constraints_to_observed_rates}
\alias{adjust_constraints_to_observed_rates}
\title{Adjust the constraint of the model to observed rates}
\usage{
adjust_constraints_to_observed_rates(model, metabolites_with_rates,
  exchange_met = build_exchange_met(model), backward_fluxes = "_b",
  forward_fluxes = "_f")
}
\arguments{
\item{model}{a genome-scale metabolic model of class modelorg}

\item{metabolites_with_rates}{is a data.frame consisting of the name of the
metabolites, their concentrations and rates in mmol/gDW/h.
The column name must be "name", "concentrations","rates"}

\item{exchange_met}{Optional. a data.frame as given by build_exchange_met}

\item{backward_fluxes}{Optional. Useful for irreversible model}

\item{forward_fluxes}{Optional. Useful for irreversible model}
}
\value{
Return the model with updated bounds corresponding to the observed
rates provided
}
\description{
Adjust the constraint of the model to observed rates
}
\examples{
data("iMM904")
metabolites_rates<-data.frame(name=c("D-Glucose","Glycerol"),
                         concentrations=c(16,0),rates=c(-2.81,-8.01))
iMM904_adjusted<-adjust_constraints_to_observed_rates(iMM904,
metabolites_rates)
}
