% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aliases_SC}
\alias{aliases_SC}
\title{aliases_SC
A data.frame containing the gene ID used in the metabolic model and their
common name, used in the gene regulatory network}
\format{a two colums data.frame which first columns correspond to the name
used in the model and the second to the ID used in the GRN (common name).
Those columns should be named geneName_model and geneName_GRN respectively.}
\usage{
aliases_SC
}
\description{
aliases_SC
A data.frame containing the gene ID used in the metabolic model and their
common name, used in the gene regulatory network
}
\keyword{datasets}
