% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicModelFunctions.R
\name{build_exchange_met}
\alias{build_exchange_met}
\title{Build the exchange metabolite data.frame}
\usage{
build_exchange_met(model)
}
\arguments{
\item{model}{An object of class modelOrg, the genome scale metabolic model}
}
\value{
a data.frame containing the exchange metabolite model id and the
equivalent name
}
\description{
Build the exchange metabolite data.frame
}
\examples{
data("iMM904")
exchanged_met<-build_exchange_met(iMM904)
head(exchanged_met)
}
