% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoregfluxFunctions.R
\name{continuous_gpr}
\alias{continuous_gpr}
\title{Continuous evaluation of the gpr rules}
\usage{
continuous_gpr(model, expression, scale = FALSE)
}
\arguments{
\item{model}{A genome-scale metabolic model of class modelorg}

\item{expression}{A numerical matrix corresponding to the gene expression.
Rownames should contain gene names/ids while samples should be in columns.}

\item{scale}{wether to scale the gene expression to unit variance}
}
\value{
A metabolic model with lower and upper bound corresponding to the
continuous version of the rule evaluation and zero on unaffected fluxes.
}
\description{
This function was adapted from code present in \cite{exp2flux}.
 Unlike exp2flux, we consider NA gene expression as uninformative, thus terms
 involving these items will not be evaluated. Also, unlike exp2flux, our
 equivalence is OR(A,B) <- sum(A,B); AND(A,B) <- min(A,B),
 more in accordance with \cite{tiger}
}
\keyword{internal}
