% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoregfluxFunctions.R
\name{coregflux_static}
\alias{coregflux_static}
\title{Update the model using the provided gene regulatory network and expression}
\usage{
coregflux_static(model, predicted_gene_expression, gene_parameter = 0,
  tol = 1e-10, aliases = NULL)
}
\arguments{
\item{model}{A genome-scale metabolic model of class modelorg}

\item{predicted_gene_expression}{The vector of predicted gene expression for
the genes present in the metabolic model as given by
\code{predict_linear_model_influence()}}

\item{gene_parameter}{Parameter of the softplus function}

\item{tol}{Fluxes values below this threshold will be ignored.}

\item{aliases}{a data.frame containing the gene names currently used in the
network under the colname "geneName" and the alias under the colnames
"alias"}
}
\value{
list containing: \item{model}{the metabolic model with the coregflux
constraints added}
\item{softplus_positive}{the results of evaluating ln(1+exp(gpr(x+theta)))
 where gpr() are the continuous version of the gpr rules applied to a set of
 gene expression x}
\item{softplus_negative}{the results of evaluating ln(1+exp(gpr(x+theta)))
 where gpr() are the continuous version of the gpr rules applied to a set of
 gene expression x}
}
\description{
\code{coregflux_static()} uses the gene states to update the fluxes bounds
from the metabolic model.
}
\examples{
data("SC_GRN_1")
data("SC_experiment_influence")
data("SC_EXP_DATA")
data("aliases_SC")
data(iMM904)
data(PredictedGeneState)
static_list<-coregflux_static(iMM904,PredictedGeneState)
}
