% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFBAsimulationFunctions.R
\name{euler_step_biomass}
\alias{euler_step_biomass}
\title{Euler step biomass}
\usage{
euler_step_biomass(biomass_t0, rate, time_step)
}
\arguments{
\item{biomass_t0}{biomass in the current time step}

\item{rate}{the growth rate for exponential growth}

\item{time_step}{time step given by t1-t0}
}
\value{
biomass after time.step=t1-t0: biomass.t1 = biomass.t0 *
exp(rate * time.step)
}
\description{
This function updates the biomass given the biomass at the
previous time step and the size of the time step.
}
\seealso{
euler_step_metabolites, update_system_state
}
\keyword{internal}
