% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFBAsimulationFunctions.R
\name{euler_step_metabolites}
\alias{euler_step_metabolites}
\title{Compute the metabolite concentrations at the next time point}
\usage{
euler_step_metabolites(met_concentrations_t0, fluxes, biomass_t0, rate,
  time_step)
}
\arguments{
\item{met_concentrations_t0}{metabolites concentration in the current time
step}

\item{fluxes}{fluxes states in the current time step for the exchange
metabolites}

\item{biomass_t0}{biomass in the current time step}

\item{rate}{the growth rate for exponential growth}

\item{time_step}{time step given by t1-t0}
}
\value{
metabolites concentrations at the next time point
}
\description{
Compute the metabolite concentrations at the next time point
}
\seealso{
euler_step_biomass, update_system_state
}
\keyword{internal}
