% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFBAsimulationFunctions.R
\name{FBA_step}
\alias{FBA_step}
\title{Flux balance analysis solution for the model}
\usage{
FBA_step(model)
}
\arguments{
\item{model}{An object of class modelOrg, the metabolic model.}
}
\value{
list with the element
fluxes - fluxes of the flux balance analysis solution
obj- value of the objective function for the flux balance analyis solution.
}
\description{
Here we just perform a flux balance analysis solution and store the objective
 and the fluxes
}
\keyword{internal}
