% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicModelFunctions.R
\name{get_biomass_flux_position}
\alias{get_biomass_flux_position}
\title{Get biomass flux position}
\usage{
get_biomass_flux_position(model, biomass_reaction_id = "biomass",
  biomass_reaction_name = NULL)
}
\arguments{
\item{model}{An object of class modelOrg, the genome scale metabolic model}

\item{biomass_reaction_id}{Default value "biomass"}

\item{biomass_reaction_name}{Optional, the react_name in the modelOrg under
which the biomass function can be found, such as "growth"}
}
\value{
the position of the biomass generating reaction according the the
objective
in our case we had the biomass reactions for models iMM904 and iTO977
}
\description{
Get biomass flux position
}
\examples{
data("iMM904")
get_biomass_flux_position(iMM904)
}
