% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FVAFunctions.R
\name{get_fva_intervals_from_observations}
\alias{get_fva_intervals_from_observations}
\title{Get intervals of flux variability (FVA) from an observed growth rate}
\usage{
get_fva_intervals_from_observations(model, observed_growth_rate,
  metabolites_rates = NULL,
  biomass_flux_index = get_biomass_flux_position(model))
}
\arguments{
\item{model}{a genome-scale metabolic model of class modelorg}

\item{observed_growth_rate}{a numerical value for the observed growth rate}

\item{metabolites_rates}{Optional. a data.frame consisting of the name of the
metabolites, their concentrations and rates in mmol/gDW/h to adjust the model
uptake rates. The column name must be "name", "concentrations","rates"}

\item{biomass_flux_index}{Optional. Index of the biomass flux as returned by
\code{get_biomass_flux_position()}}
}
\value{
Return the interval of fluxes values compatible with the observed
growth rate through flux variability analysis
}
\description{
Get intervals of flux variability (FVA) from an observed growth rate
}
\examples{
data("iMM904")
metabolites_rates<-data.frame(name=c("D-Glucose","Glycerol"),
                         concentrations=c(16,0),rates=c(-2.81,-8.01))

FluxesVarFromObs<-get_fva_intervals_from_observations(iMM904,0.205,
metabolites_rates=metabolites_rates)
}
