% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoregfluxFunctions.R
\name{get_linear_model}
\alias{get_linear_model}
\title{Train a linear model}
\usage{
get_linear_model(train_expression,
  train_influence = regulatorInfluence(network, train_expression, minTarg
  = 10), network)
}
\arguments{
\item{train_expression}{Gene expression of the training data set, not
necessary if train_influence is supplied. Should be numerical matrix
corresponding to the gene expression. Rownames should contain gene names/ids
 while samples should be in columns.}

\item{train_influence}{Optional. Regulator influence scores computed using
the function CoRegNet::regulatorInfluence for the training data set,
default minTarg = 10}

\item{network}{CoRegNet object use to build the linear model and to compute
the influence.}
}
\value{
A linear model
}
\description{
Here we train a linear regression model of the form x= alpha + beta*I
where x is the gene expression of the metabolic genes of the train data set
train_expression, alpha is an intercept, I is the influence of the regulators
 of the training data set and beta are the coefficients.
}
\details{
train_expression Gene expression of the training data set, not
 necessary if train_influence is supplied. Should be numerical matrix
 corresponding to the gene expression. Rownames should contain gene names/ids
  while samples should be in columns.
}
\seealso{
predict_linear_model_influence
}
