% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolicModelFunctions.R
\name{get_metabolites_exchange_fluxes}
\alias{get_metabolites_exchange_fluxes}
\title{Get metabolites exchange fluxes}
\usage{
get_metabolites_exchange_fluxes(model, metabolites,
  exchange_met = build_exchange_met(model), backward_fluxes = "_b",
  forward_fluxes = "_f")
}
\arguments{
\item{model}{An object of class modelOrg, the genome scale metabolic model}

\item{metabolites}{A data.frame containing the names and concentrations of
metabolites}

\item{exchange_met}{A data.frame as build by the function build_exchange_met}

\item{backward_fluxes}{Optional parameter for irreversible model to indicate
backward fluxes}

\item{forward_fluxes}{Optional parameter for irreversible model to indicate
forward fluxes}
}
\value{
a data.frame containing the exchange metabolite model id and the
equivalent name
}
\description{
Get metabolites exchange fluxes
}
\examples{
data("iMM904")
metabolites<-data.frame("name"=c("D-Glucose","Glycerol"),
                        "concentrations"=c(16,0))
get_metabolites_exchange_fluxes(iMM904,metabolites)
}
