% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoregfluxFunctions.R
\name{gpr_expression}
\alias{gpr_expression}
\title{Transform a logic gpr rule string to  a list and then evaluate the
continuous version of the rules}
\usage{
gpr_expression(gpr, expression)
}
\arguments{
\item{gpr}{A gene-protein rules from the genome-scale metabolic model}

\item{expression}{A numerical matrix corresponding to the gene expression.
Rownames should contain gene names/ids while samples should be in columns.}
}
\value{
A numerical value corresponding to the evaluation of the flux based
on the expression of the metabolic genes present in the gpr
}
\description{
Transform a logic gpr rule string to  a list and then evaluate the
continuous version of the rules
}
\keyword{internal}
