% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iMM904}
\alias{iMM904}
\title{iMM904}
\format{a modelOrg object as required by sybil. See sybilSBML for more
information on how to load other model.}
\usage{
iMM904
}
\description{
A S. cerevisiae genome-scale metabolic model as a modelOrg object
}
\keyword{datasets}
