% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODToGenes.R
\name{ODCurveToFluxCurves}
\alias{ODCurveToFluxCurves}
\title{ODCurveToFluxCurves}
\usage{
ODCurveToFluxCurves(model, ODs, times, metabolites_rates = NULL,
  biomass_flux_index = get_biomass_flux_position(model))
}
\arguments{
\item{model}{An object of class modelOrg, the metabolic model.}

\item{ODs}{A vector of measured ODs}

\item{times}{A vector of timepoints at which the flux balance analysis
solution will be evaluated.}

\item{metabolites_rates}{A data.frame containing the extraneous metabolites,
their initial concentrations and their uptake rates. Columns must be named
"names","concentrations" and "rates".}

\item{biomass_flux_index}{Optional. index of the flux corresponding to the
biomass reaction.}
}
\value{
An object FluxCurves to visualize using the function
\code{visFluxCurves}
}
\description{
This function takes measured ODs and turn them into a FluxCurves object to be
visualize using \code{visFluxCurves()}. It relies on flux variability analysis
to highlight the flux value interval required to meet the specified OD.
}
\examples{
data(iMM904)
ODs<-seq.int(0.099,1.8,length.out = 5)
times = seq(0.5,2,by=0.5)

metabolites_rates <- data.frame("name"=c("D-Glucose"),
"concentrations"=c(16.6),"rates"=c(-2.81))

ODtoflux<-ODCurveToFluxCurves(model = iMM904,
ODs = ODs,times = times, metabolites_rates = metabolites_rates)

visFluxCurves(ODtoflux)
}
\seealso{
visFluxCurves, ODCurveToMetabolicGeneCurves,
visMetabolicGeneCurves
}
