% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODToGenes.R
\name{ODCurveToMetabolicGeneCurves}
\alias{ODCurveToMetabolicGeneCurves}
\title{ODCurveToMetabolicGeneCurves}
\usage{
ODCurveToMetabolicGeneCurves(times, ODs, metabolites_rates = NULL, model,
  softplusParam = 0, singlePointFluxEstimate = FALSE,
  biomass_flux_index = get_biomass_flux_position(model),
  aliases = NULL)
}
\arguments{
\item{times}{A vector of timepoints at which the flux balance analysis
solution will be evaluated.}

\item{ODs}{vector of measured ODs.}

\item{metabolites_rates}{A data.frame containing the extraneous metabolites,
their initial concentrations and their uptake rates. Columns must be named
"names","concentrations" and "rates".}

\item{model}{An object of class modelOrg, the metabolic model.}

\item{softplusParam}{Softplus parameter identify through calibration.}

\item{singlePointFluxEstimate}{Optional, logical.}

\item{biomass_flux_index}{index of the flux corresponding to the biomass
reaction.}

\item{aliases}{Optional. A data.frame containing the gene names used in the
metabolic model and the aliases to use to match the regulatory network.}
}
\value{
Metabolic genes curves to visualize using the function
\code{visMetabolicGeneCurves}
}
\description{
This function takes measured ODs and turn them into a ODcurveToMetCurve object
to be visualize using \code{visMetabolicGeneCurves()}. It relies on flux
variability analysis to highlight the flux value interval required to meet the
specified OD and to map it on the metabolic genes.
}
\examples{
ODs<-c(0.4500000,0.5322392,0.6295079,0.7445529)
data("aliases_SC","iMM904")
ODcurveToMetCurve<-ODCurveToMetabolicGeneCurves(times = seq(0.5,2,by=0.5),
ODs = ODs,model = iMM904,aliases = aliases_SC)
visMetabolicGeneCurves(ODcurveToMetCurve,genes="YJR077C")
}
