% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ODcurveToMetCurve}
\alias{ODcurveToMetCurve}
\title{ODcurveToMetCurve
List as obtained by the function ODcurveToMetCurve}
\format{List as obtained by the function ODcurveToMetCurve}
\usage{
ODcurveToMetCurve
}
\description{
ODcurveToMetCurve
List as obtained by the function ODcurveToMetCurve
}
\keyword{datasets}
