% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODToGenes.R
\name{ODToFluxBounds}
\alias{ODToFluxBounds}
\title{ODToFluxBounds}
\usage{
ODToFluxBounds(odRate, model, metabolites_rates = NULL,
  biomass_flux_index = get_biomass_flux_position(model))
}
\arguments{
\item{odRate}{The values of OD measured over time}

\item{model}{An object of class modelOrg, the metabolic model.}

\item{metabolites_rates}{A data.frame containing the extraneous metabolites,
their initial concentrations and their uptake rates. Columns must be named
"names","concentrations" and "rates".}

\item{biomass_flux_index}{Optional. index of the flux corresponding to the
biomass reaction.}
}
\value{
Flux bounds from OD
}
\description{
ODToFluxBounds
}
