% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFBAsimulationFunctions.R
\name{perturbation_function}
\alias{perturbation_function}
\title{Perturbation function}
\usage{
perturbation_function(old_bound, tf_influence, expression_factor = 1)
}
\arguments{
\item{old_bound}{Current value of the flux bound}

\item{tf_influence}{Influence of the chosen TF in the studied condition}

\item{expression_factor}{Numerical value by which the current value of
the bounds can be multiply with}
}
\value{
New bound value
}
\description{
Perturbation function
}
\keyword{internal}
