% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PredictedGeneState}
\alias{PredictedGeneState}
\title{PredictedGeneState}
\format{a named vector}
\usage{
PredictedGeneState
}
\description{
Predicted gene state as obtained by the function
predict_linear_model_influence
}
\keyword{datasets}
