% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SC_experiment_influence}
\alias{SC_experiment_influence}
\title{SC_experiment_influence data}
\format{a named numerical vector}
\usage{
SC_experiment_influence
}
\description{
A vector of influence computed from the first sample of SC_Test_data
}
\keyword{datasets}
