% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoregfluxFunctions.R
\name{train_continuous_model}
\alias{train_continuous_model}
\title{Train a linear model and predict the gene expression from an experiment
influence}
\usage{
train_continuous_model(train_expression, train_influence, minTarget = 10,
  experiment_influence, network)
}
\arguments{
\item{train_expression}{Gene expression of the training data set, not
necessary if train_influence is supplied. Should be numerical matrix
corresponding to the gene expression. Rownames should contain gene names/ids
 while samples should be in columns.}

\item{train_influence}{Regulator influence scores of the train data set.}

\item{minTarget}{The minimum number of targets for a regulator to be
considered for actvity prediction when computing the influence.
Default set to 10}

\item{experiment_influence}{Regulator influence scores for the condition of
interest as a named vector with the TF as names.}

\item{network}{CoRegNet object to be interrogated for building the linear
models}
}
\value{
The predicted gene expression levels compute from the linear model
and the experiment influence
}
\description{
Train a linear model and predict the gene expression from an experiment
influence
}
\keyword{internal}
