% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFBAsimulationFunctions.R
\name{update_fluxes_constraints_geneKOOV}
\alias{update_fluxes_constraints_geneKOOV}
\title{Update the fluxes constraints to simulate gene KO or overexpression}
\usage{
update_fluxes_constraints_geneKOOV(model, gene_table, aliases = NULL)
}
\arguments{
\item{model}{An object of class modelOrg, the metabolic model.}

\item{gene_table}{A data.frame containing 2 columns: "gene",
"expression" containing respectively the name of a gene present in
the CoRegNet object as string and an expression factor of 0 for a KO,
or an integer >1 for an overexpression}

\item{aliases}{Optional. A data.frame containing the gene names used in the
metabolic model and the aliases to use to match the regulatory network}
}
\value{
Return the model with updated bounds
}
\description{
Update the constraints of the reactions associated with the knock-out or
overexpressed gene
}
\examples{
data("iMM904")
data("aliases_SC")
gene_table <- data.frame("gene" = c("YGL202W","YIL162W"),
"expression" =c(2,0), stringsAsFactors = FALSE)

model_gene_KO_OV_constraints <- update_fluxes_constraints_geneKOOV(
model= iMM904,
gene_table =  gene_table,
aliases = aliases_SC)
}
