% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFBAsimulationFunctions.R
\name{update_fluxes_constraints_GRegulation}
\alias{update_fluxes_constraints_GRegulation}
\title{Update fluxes constraints}
\usage{
update_fluxes_constraints_GRegulation(model, target_genes_names,
  p_function)
}
\arguments{
\item{model}{An object of class modelOrg, the metabolic model.}

\item{target_genes_names}{Targets which will be affected by the knock-out}

\item{p_function}{Perturbation function}
}
\value{
Return the model with updated bounds
}
\description{
Update the fluxes constraints to simulate a knockout/over-expression using a
penalty function (logistic) over the targets of a given TF
}
\keyword{internal}
