% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFBAsimulationFunctions.R
\name{update_fluxes_constraints_influence}
\alias{update_fluxes_constraints_influence}
\title{Update the fluxes constraints to simulate TF KO or overexpression}
\usage{
update_fluxes_constraints_influence(model, coregnet, regulator_table,
  aliases)
}
\arguments{
\item{model}{An object of class modelOrg, the metabolic model.}

\item{coregnet}{Object of class CoRegNet, containing the regulatory and
coregulatory interactions.}

\item{regulator_table}{A data.frame containing 3 columns: "regulator",
"influence","expression" containing respectively the name of a TF present in
the CoRegNet object as string, its influence in the condition of interest as
a numerical and an expression factor of 0 for a KO, or an integer >1 for an
 overexpression}

\item{aliases}{Optional, a data.frame containing the gene names used in the
metabolic model and the aliases to use to match the regulatory network}
}
\value{
Return the model with updated bounds
}
\description{
Update the constraints according to the influence & regulatory network for a
single KO or over-expression
}
\examples{
data("SC_GRN_1")
data("iMM904")
data("aliases_SC")
regulator_table <- data.frame("regulator" = "MET32",
                              "influence" = -1.20322 ,
                              "expression" = 3,
                              stringsAsFactors = FALSE)
model_TF_KO_OV_constraints <- update_fluxes_constraints_influence(
model= iMM904,coregnet = SC_GRN_1,regulator_table =  regulator_table,
aliases = aliases_SC)
}
