% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFBAsimulationFunctions.R
\name{update_fluxes_state}
\alias{update_fluxes_state}
\title{Update the fluxes bounds given the metabolites concentrations}
\usage{
update_fluxes_state(model, met_fluxes_indexes, biomass_flux_index)
}
\arguments{
\item{model}{An object of class modelOrg, the metabolic model.}

\item{met_fluxes_indexes}{index of the fluxes corresponding to uptake of
extrogenous metabolites}

\item{biomass_flux_index}{index of the flux corresponding to the biomass
reaction.}
}
\value{
a list of elements:
fluxes : array containing the fluxes values for the fba solution.
biomass_yield:  flux value for the corresponding biomass reaction
met_fluxes:  fluxes corresponding to the uptake reactions of the extraneous
metabolites
obj: value of the objective function obtained by the flux balance analysis
model.
}
\description{
Updates the fluxes values by performing flux balance analysis, it is
necessary to indicate the indexes of the extraneous metabolites and the
biomass reaction. It performs an fba step and selects the given fluxes
}
\keyword{internal}
