% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFBAsimulationFunctions.R
\name{update_system_state}
\alias{update_system_state}
\title{Update the fluxes, metabolites concentration and biomass in the system}
\usage{
update_system_state(flux_state, biomass_t0, met_concentrations_t0,
  time_step, Biomass_Yield_To_Rate)
}
\arguments{
\item{flux_state}{Current fluxes states}

\item{biomass_t0}{Initial biomass}

\item{met_concentrations_t0}{Metabolites concentration at t0}

\item{time_step}{Time step}

\item{Biomass_Yield_To_Rate}{Function to convert biomass to rate.
Default is f(x)=1*x}
}
\value{
Return updated biomass and metabolites concentrations
}
\description{
Update the fluxes, metabolites concentration and biomass in the system
}
\keyword{internal}
