% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFBAsimulationFunctions.R
\name{update_uptake_fluxes_constraints_metabolites}
\alias{update_uptake_fluxes_constraints_metabolites}
\title{Update the fluxes constraints given the metabolite concentrations}
\usage{
update_uptake_fluxes_constraints_metabolites(model, met_fluxes_indexes,
  met_concentrations_t0, biomass_t0, time_step)
}
\arguments{
\item{model}{An object of class modelOrg, the metabolic model.}

\item{met_fluxes_indexes}{Indexes of the metabolites fluxes}

\item{met_concentrations_t0}{Metabolites concentrations at t0}

\item{biomass_t0}{Biomasss at t0}

\item{time_step}{time_step studied}
}
\value{
Return the updated model
}
\description{
Update the fluxes constraints given the metabolite concentrations
}
