% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODToGenes.R
\name{visFluxCurves}
\alias{visFluxCurves}
\title{Visualize Fluxes Curves}
\usage{
visFluxCurves(fluxCurves, genes = unique(fluxCurves$name)[1:50], ...)
}
\arguments{
\item{fluxCurves}{result table from ODCurveToFluxCurves}

\item{genes}{a vector containing the names of the metabolic genes to plot.
Default select the first 50 genes}

\item{...}{Optional others curves}
}
\value{
a plot of the curves of the chosen fluxes
}
\description{
Visualize Fluxes Curves
}
\examples{
ODs<-seq.int(0.099,1.8,length.out = 5)
times = seq(0.5,2,by=0.5)

metabolites_rates <- data.frame("name"=c("D-Glucose"),
"concentrations"=c(16.6),"rates"=c(-2.81))

ODtoflux<-ODCurveToFluxCurves(model = iMM904,
ODs = ODs,times = times, metabolites_rates = metabolites_rates)

visFluxCurves(ODtoflux,genes ="ADK3")
}
\seealso{
ODCurveToFluxCurves, ODCurveToMetabolicGeneCurves,
visMetabolicGeneCurves
}
