% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODToGenes.R
\name{visMetabolicGeneCurves}
\alias{visMetabolicGeneCurves}
\title{Visualize Metabolic Gene Curves}
\usage{
visMetabolicGeneCurves(metabCurves,
  genes = unique(metabCurves$name)[1:50], ...)
}
\arguments{
\item{metabCurves}{result table from ODCurveToMetabolicGeneCurves}

\item{genes}{a vector containing the names of the metabolic genes to plot.
Default select the first 50 genes}

\item{...}{Optional others curves}
}
\value{
a plot of the curves of the chosen metabolic genes
}
\description{
Visualize Metabolic Gene Curves
}
\examples{
ODs<-seq.int(0.099,1.8,length.out = 5)
times = seq(0.5,2,by=0.5)
metabolites_rates <- data.frame("name"=c("D-Glucose"),
"concentrations"=c(16.6),"rates"=c(-2.81))

ODcurveToMetCurve<- ODCurveToMetabolicGeneCurves(times = times,
ODs = ODs,model = iMM904,aliases = aliases_SC,
metabolites_rates = metabolites_rates)

visMetabolicGeneCurves(ODcurveToMetCurve,genes="YJR077C")
}
\seealso{
ODCurveToMetabolicGeneCurves,ODCurveToFluxCurves, visFluxCurves
}
