% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ODcurveToMetCurve}
\alias{ODcurveToMetCurve}
\title{ODcurveToMetCurve data}
\format{List as obtained by the function \code{ODCurveToMetabolicGeneCurves}}
\usage{
ODcurveToMetCurve
}
\description{
List as obtained by the function \code{ODCurveToMetabolicGeneCurves}
}
\keyword{datasets}
