% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ODtoflux}
\alias{ODtoflux}
\title{ODtoflux data}
\format{List as obtained by the function \code{ODCurveToFluxCurves}}
\usage{
ODtoflux
}
\description{
List as obtained by the function \code{ODCurveToFluxCurves}
}
\keyword{datasets}
