% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PredictedGeneState}
\alias{PredictedGeneState}
\title{PredictedGeneState data}
\format{a named vector containing the gene name and its associated predicted
gene state.}
\usage{
PredictedGeneState
}
\description{
Predicted gene states as obtained by the function
\code{predict_linear_model_influence}
}
\keyword{datasets}
