% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SC_EXP_DATA}
\alias{SC_EXP_DATA}
\title{SC_EXP_DATA data}
\format{a matrix of 3600 genes by 247 samples}
\source{
subset of m3d dataset available at <http://m3d.mssm.edu/>
}
\usage{
SC_EXP_DATA
}
\description{
A matrix of S. cerevisiae gene expression in various experimental designs,
derived from the m3d dataset to infer _S. cerevisiae_ gene regulatory network.
The dataset was shorten to 3600 genes in order to limit the size of the
object
}
\keyword{datasets}
