% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SC_GRN_1}
\alias{SC_GRN_1}
\title{SC_GRN_1 data}
\format{a coregnet object inferred using the package _CoRegNet_
\describe{
  \item{Number of transcription factor }{200}
  \item{Number of targets genes}{3748}
  \item{Evidences}{TRUE}
  ...
}}
\usage{
SC_GRN_1
}
\description{
A coregnet object infered from the m3d dataset describing the gene
regulatory network for S. cerevisiae as described in Banos, D. T., Trébulle,
P., & Elati, M. (2017). Integrating transcriptional activity in genome-scale
models of metabolism. BMC systems biology, 11(7), 134.
}
\keyword{datasets}
