% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SC_Test_data}
\alias{SC_Test_data}
\title{SC_Test_data data}
\format{a matrix of 6028 genes by 13 samples during diauxic shift}
\source{
E-GEOD-4398 (Brauer MJ and al.)
}
\usage{
SC_Test_data
}
\description{
A matrix of S. cerevisiae gene expression during diauxic shift (Brauer and
al.)
}
\keyword{datasets}
