% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODToGenes.R
\name{visFluxCurves}
\alias{visFluxCurves}
\title{Visualize Fluxes Curves}
\usage{
visFluxCurves(fluxCurves, genes = unique(fluxCurves$name)[seq_len(50)],
  ...)
}
\arguments{
\item{fluxCurves}{result table from ODCurveToFluxCurves}

\item{genes}{a vector containing the names of the metabolic genes to plot.
Default select the first 50 genes}

\item{...}{Optional others curves}
}
\value{
a plot of the curves of the chosen fluxes
}
\description{
Visualize Fluxes Curves
}
\examples{
data("ODtoflux")
visFluxCurves(ODtoflux,genes ="ADK3")
}
\seealso{
ODCurveToFluxCurves, ODCurveToMetabolicGeneCurves,
visMetabolicGeneCurves
}
