% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODToGenes.R
\name{visMetabolicGeneCurves}
\alias{visMetabolicGeneCurves}
\title{Visualize Metabolic Gene Curves}
\usage{
visMetabolicGeneCurves(metabCurves,
  genes = unique(metabCurves$name)[seq_len(50)], ...)
}
\arguments{
\item{metabCurves}{result table from ODCurveToMetabolicGeneCurves}

\item{genes}{a vector containing the names of the metabolic genes to plot.
Default select the first 50 genes}

\item{...}{Optional, others curves}
}
\value{
a plot of the curves of the chosen metabolic genes
}
\description{
Visualize Metabolic Gene Curves
}
\examples{
data("ODcurveToMetCurve")

visMetabolicGeneCurves(ODcurveToMetCurve,genes="YJR077C")
}
\seealso{
ODCurveToMetabolicGeneCurves,ODCurveToFluxCurves, visFluxCurves
}
